rag-pipeline/requirements.txt
python-dotenv==1.0.1
qdrant-client==1.12.1
google-generativeai==0.8.3
markdown==3.7
beautifulsoup4==4.12.3
PyYAML==6.0.2
